fix: Don't set control_id to null in create, let boot() auto-generate it
When control_id was explicitly set to null, Eloquent tried to insert

NULL which violated the NOT NULL constraint. The boot() method's

creating event couldn't fix this because the attribute was already set.

Solution: Only include control_id in create() attributes if it's

actually provided in the request data. Otherwise, let the model's

boot() method auto-generate it via the creating event.